+++ stonks.h
@@ -26,8 +26,8 @@ extern struct state
 extern struct state* initState();
 void updateInstrument(struct state*, int, float);
 void updateInstruments(struct state*);
-void sellInstrument(struct state*, int, int);
-void buyInstrument(struct state*, int, int);
+void sellInstrument(struct state*, int, float);
+void buyInstrument(struct state*, int, float);
 float portfolioValue(struct state*);
 
 #endif
